Fix door sensor on H2D and newer X1
As part of the work for H2D support, Bambu Lab moved the door open
flag from 'home_flag' to 'stat'. Add a dataclass to support parsing
stat, and update the door open sensor to use stat when this object
indicates door open is available.

The X1 series firmware version that moved this flag is unknown, but is
presumed to be 1.09 based on other major updates including the user
interface, support for AMS 2 Pro/AMS HT, and the recommendation to
update Bambu Studio to 2.0.3.54.

Fixes #1421.