Fix door sensor on H2D and newer X1
As part of the work for H2D support, Bambu Lab moved the door open
flag from 'home_flag' to 'stat'. Move the door status to the Info
class, and update Info to parse stat instead of home_flag on newer
printers and X1 series with newer firmware.

The X1 series firmware version that moved this flag is presumed to
be 1.09 based on other major updates including the user interface,
support for AMS 2 Pro/AMS HT, and the recommendation to update
Bambu Studio to 2.0.3.54.

Fixes #1421.

Change-Id: I45a16275d58a80bb45c3d2074ac73b0cf99859f6