Fix door sensor on X1 series
The assumption that the door sensor moved to the 'stat' flag for newer
X1 and X1C has been shown to be inaccurate. Revert the behavior change
from #1436 for these printers, so that 'stat' is only used for the H2
series of printers and 'home_flag' is used for X1/X1C.

Fixes #1548.