Add dual-nozzle sizes and types
Add sensors for the left/right nozzle size (diameter) and type,
parsed from the device.nozzle.info[] array using the same code
pattern as device.extruder parsing. Consistent with the nozzle
temperature sensors, the nozzle_diameter and nozzle_type sensor
values are that of the active nozzle. "Active" is used in
pybambu to clarify this, but the existing sensors are not renamed
to preserve backward compatibility.

This fixes the nozzle size and type for H2D printers, and adds
tungsten_carbide and high_flow nozzle types.