Change 'high flow' to 'with large flow' for translation
Edit the English descriptions for high-flow nozzle types from e.g.
"high flow stainless steel" to "stainless steel with large flow".
Moving away from idiomatic English toward a more technically precise
form is intended to elicit the correct terms and declensions. For
example, 'high' should not be interpreted as an altitude, and the
'flow' is not of the nozzle material itself.