Improve nozzle type and diameter support for H2D (#1551)
* Add dual-nozzle sizes and types

Add sensors for the left/right nozzle size (diameter) and type,
parsed from the device.nozzle.info[] array using the same code
pattern as device.extruder parsing. Consistent with the nozzle
temperature sensors, the nozzle_diameter and nozzle_type sensor
values are that of the active nozzle. "Active" is used in
pybambu to clarify this, but the existing sensors are not renamed
to preserve backward compatibility.

This fixes the nozzle size and type for H2D printers, and adds
tungsten_carbide and high_flow nozzle types.

* Change 'high flow' to 'with large flow' for translation

Edit the English descriptions for high-flow nozzle types from e.g.
"high flow stainless steel" to "stainless steel with large flow".
Moving away from idiomatic English toward a more technically precise
form is intended to elicit the correct terms and declensions. For
example, 'high' should not be interpreted as an altitude, and the
'flow' is not of the nozzle material itself.

* Update translations

Add translations for new strings, generated with auto_translate.py.

* Revert "Change 'high flow' to 'with large flow' for translation"

This reverts commit 81c940a1bd639704bbba41ee51655afeb798d248. The
idiomatic form is desired for English.

* Update doc

Clarify that H2D has three entities for each nozzle attribute, and
that the entities without left/right are for the active nozzle.